+++ filament/src/Material.cpp
@@ -298,6 +298,15 @@ bool FMaterial::hasParameter(const char* name) const noexcept {
     return true;
 }
 
+UniformInterfaceBlock::UniformInfo const* FMaterial::reflect(
+        utils::StaticString const& name) const noexcept {
+    auto const& list = mUniformInterfaceBlock.getUniformInfoList();
+    auto p = std::find_if(list.begin(), list.end(), [&](auto const& e) {
+        return e.name == name;
+    });
+    return p == list.end() ? nullptr : &static_cast<UniformInterfaceBlock::UniformInfo const&>(*p);
+}
+
 backend::Handle<backend::HwProgram> FMaterial::getProgramSlow(uint8_t variantKey) const noexcept {
     switch (getMaterialDomain()) {
         case MaterialDomain::SURFACE: