+++ filament/src/details/MaterialInstance.h
@@ -61,10 +61,10 @@ public:
         }
     }
 
-    template <typename T>
+    template <typename T, typename = is_supported_parameter_t<T>>
     void setParameter(const char* name, T value) noexcept;
 
-    template <typename T>
+    template <typename T, typename = is_supported_parameter_t<T>>
     void setParameter(const char* name, const T* value, size_t count) noexcept;
 
     void setParameter(const char* name,