+++ filament/src/details/Renderer.h
@@ -100,20 +100,20 @@ private:
             uint32_t xoffset, uint32_t yoffset, uint32_t width, uint32_t height,
             backend::PixelBufferDescriptor&& buffer);
 
-    static RenderPass::CommandTypeFlags getCommandType(View::DepthPrepass prepass) noexcept;
-
-    struct PrepareColorPassesData {
-        FrameGraphId<FrameGraphTexture> ssao;
-        FrameGraphId<FrameGraphTexture> color;
-        FrameGraphId<FrameGraphTexture> depth;
+    struct ColorPassConfig {
         Viewport svp;
         backend::TextureFormat hdrFormat;
         uint8_t msaa;
     };
 
-    static FrameGraphId<FrameGraphTexture> colorPass(FrameGraph& fg,
-            PrepareColorPassesData& blackboard, RenderPass const& pass,
-            backend::TargetBufferFlags clearFlags, math::float4 clearColor = {}) noexcept;
+    static FrameGraphId<FrameGraphTexture> colorPass(FrameGraph& fg, const char* name,
+            FrameGraphTexture::Descriptor const& colorBufferDesc, ColorPassConfig const& config,
+            RenderPass const& pass, backend::TargetBufferFlags clearFlags,
+            math::float4 clearColor = {}) noexcept;
+
+    FrameGraphId<FrameGraphTexture> refractionPass(FrameGraph& fg,
+            ColorPassConfig const& config, RenderPass const& pass,
+            FView const& view, backend::TargetBufferFlags clearFlags) const noexcept;
 
     void recordHighWatermark(size_t watermark) noexcept {
         mCommandsHighWatermark = std::max(mCommandsHighWatermark, watermark);