+++ ios/samples/hello-pbr/hello-pbr/FilamentApp.cpp
@@ -44,7 +44,6 @@ void FilamentApp::initialize() {
     camera = engine->createCamera();
 
     filaView = engine->createView();
-    filaView->setDepthPrepass(filament::View::DepthPrepass::DISABLED);
 
     image::KtxBundle* iblBundle = new image::KtxBundle(RESOURCES_VENETIAN_CROSSROADS_2K_IBL_DATA,
             RESOURCES_VENETIAN_CROSSROADS_2K_IBL_SIZE);