+++ libs/gltfio/include/gltfio/SimpleViewer.h
@@ -138,12 +138,6 @@ public:
     void enableDithering(bool b) { mEnableDithering = b; }
 
     /**
-     * Enables depth prepass on the view.
-     * Defaults to true.
-     */
-    void enablePrepass(bool b) { mEnablePrepass = b; }
-
-    /**
      * Enables FXAA antialiasing in the post-process pipeline.
      * Defaults to true.
      */
@@ -195,7 +189,6 @@ private:
     bool mEnableSunlight = true;
     bool mEnableShadows = true;
     bool mEnableDithering = true;
-    bool mEnablePrepass = true;
     bool mEnableFxaa = true;
     bool mEnableMsaa = true;
     bool mEnableSsao = true;
@@ -265,14 +258,20 @@ SimpleViewer::~SimpleViewer() {
 }
 
 void SimpleViewer::setAsset(FilamentAsset* asset, bool scale) {
-    using namespace filament::math;
+    if (mAsset == asset) {
+        return;
+    }
     removeAsset();
     mAsset = asset;
+    if (!asset) {
+        mAnimator = nullptr;
+        return;
+    }
     mAnimator = asset->getAnimator();
     if (scale) {
         auto& tcm = mEngine->getTransformManager();
         auto root = tcm.getInstance(mAsset->getRoot());
-        mat4f transform = fitIntoUnitCube(mAsset->getBoundingBox());
+        filament::math::mat4f transform = fitIntoUnitCube(mAsset->getBoundingBox());
         tcm.setTransform(root, transform);
     }
     mScene->addEntities(mAsset->getEntities(), mAsset->getEntityCount());
@@ -313,6 +312,9 @@ void SimpleViewer::updateIndirectLight() {
 }
 
 void SimpleViewer::applyAnimation(double currentTime) {
+    if (!mAnimator) {
+        return;
+    }
     static double startTime = 0;
     const size_t numAnimations = mAnimator->getAnimationCount();
     if (mResetAnimation) {
@@ -417,14 +419,11 @@ void SimpleViewer::updateUserInterface() {
 
     if (ImGui::CollapsingHeader("View")) {
         ImGui::Checkbox("Dithering", &mEnableDithering);
-        ImGui::Checkbox("Depth prepass", &mEnablePrepass);
         ImGui::Checkbox("FXAA", &mEnableFxaa);
         ImGui::Checkbox("MSAA 4x", &mEnableMsaa);
         ImGui::Checkbox("SSAO", &mEnableSsao);
     }
 
-    mView->setDepthPrepass(
-            mEnablePrepass ? View::DepthPrepass::ENABLED : View::DepthPrepass::DISABLED);
     mView->setDithering(mEnableDithering ? View::Dithering::TEMPORAL : View::Dithering::NONE);
     mView->setAntiAliasing(mEnableFxaa ? View::AntiAliasing::FXAA : View::AntiAliasing::NONE);
     mView->setSampleCount(mEnableMsaa ? 4 : 1);